Catch exceptions coming from decrypt()
A nuber of exceptions may be thrown by various libraries used in decrypt.

One especially important one would be that URI.parse functions do not
validate a URI, and will raise errors if they get an invalid URI.

We should catch the error and provide a sensible return value to allow
the user to know that decrypt has failed and why.